Move FinalOutcome to steps folder
